Some cleanup

+++ src/core/Codecs/Lucene40/Lucene40FieldInfosFormat.cs
@@ -31,12 +31,12 @@ namespace Lucene.Net.Codecs.Lucene40
         internal const int FORMAT_START = 0;
         internal const int FORMAT_CURRENT = FORMAT_START;
 
-        internal const byte IS_INDEXED = 0x1;
-        internal const byte STORE_TERMVECTOR = 0x2;
-        internal const byte STORE_OFFSETS_IN_POSTINGS = 0x4;
-        internal const byte OMIT_NORMS = 0x10;
-        internal const byte STORE_PAYLOADS = 0x20;
-        internal const byte OMIT_TERM_FREQ_AND_POSITIONS = 0x40;
-        internal const byte OMIT_POSITIONS = (byte)(sbyte)-128;
+        internal const sbyte IS_INDEXED = 0x1;
+        internal const sbyte STORE_TERMVECTOR = 0x2;
+        internal const sbyte STORE_OFFSETS_IN_POSTINGS = 0x4;
+        internal const sbyte OMIT_NORMS = 0x10;
+        internal const sbyte STORE_PAYLOADS = 0x20;
+        internal const sbyte OMIT_TERM_FREQ_AND_POSITIONS = 0x40;
+        internal const sbyte OMIT_POSITIONS = -128;
     }
 }
